LayoutTests/fast/mutation/mutation-record-nullity-expected.txt

Non-relevant properties on mutation records should be null

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


characterData record:
PASS record.attributeName is null
PASS record.attributeNamespace is null
PASS record.oldValue is null
PASS record.addedNodes is null
PASS record.removedNodes is null
PASS record.previousSibling is null
PASS record.nextSibling is null

childList record:
PASS record.attributeName is null
PASS record.attributeNamespace is null
PASS record.oldValue is null

attributes record:
PASS record.attributeNamespace is null
PASS record.oldValue is null
PASS record.addedNodes is null
PASS record.removedNodes is null
PASS record.previousSibling is null
PASS record.nextSibling is null
PASS successfullyParsed is true

TEST COMPLETE
